Renamed file operation tutorial properly
